Update Axes on change to chart type 'BarChart'
Changing chart type to 'BarChart' in runtime does not updates the Axes.
when going from ColumnChart/AreaChart/.. to BarChart, the vAxis should
become into hAxis and vice-versa.

To solve this problem, in my commit, when switching to BarChart and the
previous type is different from Barchart, OR when switching to any other
type and the previous type is Barchart the Axes are exchanged!